Refactor model for protocol call
